+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fix python command recognition: avoid wrapping for -c module argument
 * Support escaped space and quote during command string parsing
+* Support command definition as array/list in PythonTask 
 
 ### 1.0.0 (2017-12-20)
 * Initial release